fixed remark and moved to another folder
